break out utils and errors
